refactor(snmptraps): migrate config component to v2 (#48404)
### What does this PR do?
Migrates `comp/snmptraps/config` from v1 to v2 component structure.

- Moves the `Component` interface and `TrapsConfig` types to `config/def/`
- Moves the implementation to `config/impl/` with `Requires`/`Provides`/`NewComponent`
- Adds `config/fx/` with `fxutil.ProvideComponentConstructor`
- Updates import paths in `forwarder/forwarderimpl`, `listener/listenerimpl`, and `server/serverimpl`

### Motivation
Part of a series migrating all `comp/snmptraps` components to the v2 component framework.

### Describe how you validated your changes
Structural refactor — no logic changes. Pre-commit hooks (go-fmt, copyright) passed.

### Additional Notes
This PR is part of a chain: each PR targets the previous component's branch for a clean diff.
Chain: status → oidresolver → config → formatter → listener → forwarder → server
