+++ comp/snmptraps/config/def/config_test.go
